Large change/digression on the path to email confirmation.  Realized I would need a profile to allow the user to specify how they would like to be addressed, like "Steve" in an email etc.
So this commit implements a profile and fixes a bug with validating email where the current user's own email has to be ignored in the unique-test.

Also allows deleting an account, which I realized I would need for manual testing of registration and email confirmation.  But which is generally useful, too.